Changed Proposal Key Type

+++ src/HttpServer/Model/Mappers/ProposalMapper.cs
@@ -3,7 +3,7 @@ using HttpServer.Model.Entities;
 
 namespace HttpServer.Model.Mappers
 {
-    public class ProposalMapper : AbstractMapper<UInt32, Proposal>
+    public class ProposalMapper : AbstractMapper<long, Proposal>
     {
     }
 }
